Numbered Directories PS1
Creates numbered directories with a padding of two digits.  Asks user if they want to choose a parent directory, and takes them to a folder dialog browser to choose.  Otherwise, chooses a default location.  Asks for number of folders to be made and what number to start with for naming.