Add files via upload
Script to save image files from a series of websites.  Only works with a specific main domain, as it uses the variables from the url to iterate through the websites.  In its current state, the script would have to be customized to work with a different domain. Then the Split() function separates the strings into usable variables and the correct variables are iterated.